requirements.txt
tox
pytest
pytest-cov
pytest-pep8
coverage
coveralls
PyYAML
